+++ Tools/MyEditorEvents.cs
@@ -1,11 +1,11 @@
 #if UNITY_EDITOR
 using System;
 using JetBrains.Annotations;
+using MyBox.Internal;
 using UnityEditor;
 using UnityEditor.Build;
 using UnityEngine;
 using UnityEditor.Build.Reporting;
-using UnityEngine.SceneManagement;
 
 namespace MyBox.EditorTools
 {
@@ -22,7 +22,15 @@ namespace MyBox.EditorTools
 		/// </summary>
 		public static event Action OnFirstFrame;
 
+		/// <summary>
+		/// On Before Playmode (PlayModeStateChange is ExitingEditMode)
+		/// </summary>
 		public static event Action BeforePlaymode;
+
+		/// <summary>
+		/// On After Playmode (PlayModeStateChange is EnteredEditMode)
+		/// </summary>
+		public static event Action AfterPlaymode;
 
 		public static event Action BeforeBuild;
 
@@ -37,13 +45,30 @@ namespace MyBox.EditorTools
 		/// </summary>
 		public static event Action<Event> OnEditorInput;
 
+		/// <summary>
+		/// The same as the OnGUI call from MonoBehavior in playmode
+		/// </summary>
+		public static event Action OnPlaymodeGUI;
+
+		/// <summary>
+		/// The same as Update cal from MonoBehavior in playmode
+		/// </summary>
+		public static event Action OnBehaviourUpdate;
+
 
 		static MyEditorEvents()
 		{
 			// DelayCall is used to ensure that all [InitializeOnLoad] subscribers are initialized before the events are called 
 			EditorApplication.delayCall += () => EditorApplication.update += CheckOnceOnEditorStart;
 			EditorApplication.update += CheckOnceOnPlaymode;
+			EditorApplication.update += () =>
+			{
+				if (OnPlaymodeGUI != null) MyEditorEventsBehaviorHandler.InitializeInstance();
+				if (OnBehaviourUpdate != null) MyEditorEventsBehaviorHandler.InitializeInstance();
+			};
 
+			MyEditorEventsBehaviorHandler.OnGUIEvent += () => OnPlaymodeGUI?.Invoke();
+			MyEditorEventsBehaviorHandler.OnUpdate += () => OnBehaviourUpdate?.Invoke();
 			
 			EditorApplication.playModeStateChanged += OnPlayModeStateChanged;
 			RegisterRawInputHandler();
@@ -59,7 +84,7 @@ namespace MyBox.EditorTools
 				globalEventHandler.SetValue(null, callback);
 			}
 		}
-
+		
 		/// <summary>
 		/// On Editor Save
 		/// </summary>
@@ -99,12 +124,10 @@ namespace MyBox.EditorTools
 			}
 		}
 
-		/// <summary>
-		/// On Before Playmode
-		/// </summary>
 		private static void OnPlayModeStateChanged(PlayModeStateChange state)
 		{
-			if (state == PlayModeStateChange.ExitingEditMode && BeforePlaymode != null) BeforePlaymode();
+			if (state == PlayModeStateChange.ExitingEditMode) BeforePlaymode?.Invoke();
+			if (state == PlayModeStateChange.EnteredEditMode) AfterPlaymode?.Invoke();
 		}
 
 		private static void RawInputHandler()
@@ -114,6 +137,7 @@ namespace MyBox.EditorTools
 
 			OnEditorInput?.Invoke(e);
 		}
+		
 
 		/// <summary>
 		/// Before Build